requirements.txt
aiohttp==3.8.1
aiohttp-jinja2==1.5
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Deprecated==1.2.13
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
MarkupSafe==2.1.0
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.18.2
python-dotenv==0.19.2
redis==4.1.4
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
